Refactor Nk3Context
This patch splits the Nk3Context into two parts:  Information about
connected devices is now stored in the new DeviceData class.  The
remaining code that is only required for the update procedure is moved
into the update module.
