+++ nitrokeyapp/overview_tab.py
@@ -2,6 +2,7 @@ from typing import Optional
 
 from PyQt5.QtWidgets import QWidget
 
+from nitrokeyapp.device_data import DeviceData
 from nitrokeyapp.ui.overview_tab import Ui_OverviewTab
 
 
@@ -16,7 +17,7 @@ class OverviewTab(QWidget):
         self.ui.progressBar_Download.hide()
         self.ui.progressBar_Finalization.hide()
 
-    def refresh(self, path: str, uuid: str, version: str) -> None:
-        self.ui.nk3_lineedit_path.setText(path)
-        self.ui.nk3_lineedit_uuid.setText(uuid)
-        self.ui.nk3_lineedit_version.setText(version)
+    def refresh(self, data: DeviceData) -> None:
+        self.ui.nk3_lineedit_path.setText(str(data.path))
+        self.ui.nk3_lineedit_uuid.setText(str(data.uuid))
+        self.ui.nk3_lineedit_version.setText(str(data.version))
